+++ crates/tectonic-core/src/storage/shard.rs
@@ -10,19 +10,19 @@ use crate::storage::location::ArenaLocation;
 
 #[allow(dead_code)]
 pub struct CacheShard {
-    pub shard_id: usize,
-    pub capacity: usize,
-    pub size: usize,
+    pub shard_id: u32,
+    pub capacity: u64,
+    pub size: u64,
     pub location_storage: Vec<Option<ArenaLocation<'static>>>
 }
 
 impl CacheShard {
-    pub fn with_capacity(id: usize, capacity: usize) -> Self {
+    pub fn with_capacity(id: u32, capacity: u64) -> Self {
         Self { 
             shard_id: id,
             capacity: capacity,
             size: 0,
-            location_storage: vec![None; capacity],
+            location_storage: vec![None; capacity as usize],
         }
     }
 }
